# api/openapi/tars-mvp.yaml
openapi: 3.1.0
info:
  title: TARS MVP API
  version: 0.1.0
  description: |
    Draft API definition aligned with tars_technical_design.md v1.4.
    Tag names describe API contract groupings, not Web Console page ownership.
    In particular, the `ops` tag means authenticated control-plane APIs and is
    not equivalent to the `/ops` UI page; some `ops`-tagged APIs are primarily
    surfaced from other daily object pages such as `/identity/*`, `/providers`,
    and `/connectors`.
servers:
  - url: https://tars.example.com
...
paths:
...
  /api/v1/users/{user_id}:
    get:
      tags: [ops]
      summary: Get user
      operationId: getUser
      security:
        - OpsBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ops]
      summary: Update user
      operationId: updateUser
      security:
        - OpsBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpsertRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/connectors:
    get:
      tags: [public]
      summary: List registered connector manifests
      operationId: listConnectors
      parameters:
        - name: kind
          in: query
          schema:
            type: string
        - name: protocol
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: vendor
          in: query
          schema:
            type: string
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum:
              [
                enabled,
                id,
                name,
                display_name,
                vendor,
                kind,
                version,
                protocol,
                type,
              ]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Connector registry list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorListResponse"
    post:
      tags: [ops]
      summary: Create registered connector manifest
      operationId: createConnector
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectorUpsertRequest"
      responses:
        "201":
          description: Created connector manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectorManifest"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/StateConflict"
...
  /api/v1/skills:
    get:
      tags: [ops]
      summary: List registered skill manifests
      operationId: listSkills
      security:
        - OpsBearer: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: enabled
          in: query
          schema:
            type: boolean
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, id, version, status, source]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Skill registry list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillListResponse"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
    post:
      tags: [ops]
      summary: Create or upsert a skill manifest
      operationId: createSkill
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillUpsertRequest"
      responses:
        "201":
          description: Created skill manifest and lifecycle state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillImportResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
...
  /api/v1/extensions:
    get:
      tags: [ops]
      summary: List extension bundle candidates
      operationId: listExtensions
      security:
        - OpsBearer: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [updated_at, status, skill_id]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Extension candidate list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtensionListResponse"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
    post:
      tags: [ops]
      summary: Generate an extension bundle candidate
      operationId: createExtensionCandidate
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtensionCreateRequest"
      responses:
        "201":
          description: Generated extension candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtensionCandidate"
        "400":
          $ref: "#/components/responses/ValidationFailed"
...
  /api/v1/sessions:
    get:
      tags: [ops]
      summary: List sessions
      operationId: listSessions
      security:
        - OpsBearer: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: host
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [updated_at, created_at, status, session_id]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Session list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
...
  /api/v1/executions:
    get:
      tags: [ops]
      summary: List execution requests
      operationId: listExecutions
      security:
        - OpsBearer: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, completed_at, status, target_host, execution_id]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Execution list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionListResponse"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
...
  /api/v1/summary:
    get:
      tags: [ops]
      summary: Get dashboard summary
      operationId: getOpsSummary
      security:
        - OpsBearer: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpsSummary"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
...
  /api/v1/logs:
    get:
      tags: [ops]
      summary: List built-in runtime logs
      operationId: listRuntimeLogs
      security:
        - OpsBearer: []
      parameters:
        - name: level
          in: query
          schema:
            type: string
        - name: component
          in: query
          schema:
            type: string
        - name: session_id
          in: query
          schema:
            type: string
        - name: execution_id
          in: query
          schema:
            type: string
        - name: trace_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Runtime log list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogListResponse"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
  /api/v1/observability:
    get:
      tags: [ops]
      summary: Get built-in observability summary
      operationId: getObservabilitySummary
      security:
        - OpsBearer: []
      responses:
        "200":
          description: Observability summary, recent logs/events, and retention status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObservabilityResponse"
        "404":
          $ref: "#/components/responses/OpsAPIDisabled"
...
  /api/v1/audit:
    get:
      tags: [ops]
      summary: List audit records
      operationId: listAuditRecords
      security:
        - OpsBearer: []
      parameters:
        - name: resource_type
          in: query
          schema:
            type: string
        - name: action
          in: query
          schema:
            type: string
        - name: actor
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, resource_type, action, actor]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Audit record list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditListResponse"
...
  /api/v1/knowledge:
    get:
      tags: [ops]
      summary: List knowledge records
      operationId: listKnowledgeRecords
      security:
        - OpsBearer: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [updated_at, title, session_id]
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Knowledge record list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeListResponse"
...
components:
  securitySchemes:
    OpsBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    User:
      type: object
      properties:
        user_id:
          type: string
        username:
          type: string
        display_name:
          type: string
        email:
          type: string
        status:
          type: string
        source:
          type: string
        password_hash:
          type: string
        password_login_enabled:
          type: boolean
        password_updated_at:
          type: string
          format: date-time
        challenge_required:
          type: boolean
        mfa_enabled:
          type: boolean
        mfa_method:
          type: string
        totp_secret:
          type: string
        roles:
          type: array
          items:
            type: string
        groups:
          type: array
          items:
            type: string
        identities:
          type: array
          items:
            $ref: "#/components/schemas/IdentityLink"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UserUpsertRequest:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
        operator_reason:
          type: string
          description: Optional on IAM day-to-day user pages; backend may infer or default the audit reason.
...